[MP-4]: Update krpc version entries to 0.5.4 (#35)
* [MP-4]: Update krpc version entries to 0.5.4

* [MP-4]: Include krpc-java-0.5.4, fix double value

* [MP-4]: Bump minor version to 0.8.0

+++ src/main/java/com/pesterenan/model/ActiveVessel.java
@@ -37,7 +37,7 @@ public class ActiveVessel {
     protected Stream<Float> totalMass;
     public ReferenceFrame surfaceReferenceFrame;
     public float totalCharge;
-    public float gravityAcel;
+    public double gravityAcel;
     public CelestialBody currentBody;
     public Stream<Double> altitude;
     public Stream<Double> surfaceAltitude;
